Finished the first version of the tutorial pane - you can now select a difficulty at the end of the tutorial from a panel in the level summary state.

+++ BlammoEngine/TextLabel.h
@@ -119,8 +119,8 @@ public:
 	}
 
 	// Obtain the height of this label
-	size_t GetHeight() const {
-		return this->scale * this->font->GetHeight();
+	float GetHeight() const {
+		return this->scale * static_cast<float>(this->font->GetHeight());
 	}
 	float GetLastRasterWidth() const {
 		return this->scale * this->lastRasterWidth;
